requirements.txt
PyQt6>=6.6.0
piexif>=1.1.3
Pillow>=10.0.0
ExifRead>=3.0.0
rawpy>=0.19.0
pyinstaller